refactor(tui/attach): use dynamic import for subcommand
