+++ packages/opencode/src/index.ts
@@ -22,7 +22,7 @@ import { McpCommand } from "./cli/cmd/mcp/command"
 import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
 import { ImportCommand } from "./cli/cmd/import/command"
-import { AttachCommand } from "./cli/cmd/tui/attach"
+import { AttachCommand } from "./cli/cmd/tui/attach/command"
 import { TuiThreadCommand } from "./cli/cmd/tui/thread"
 import { AcpCommand } from "./cli/cmd/acp/command"
 import { EOL } from "os"
